% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GnnGarrotte.R
\name{G.nn.garrotte}
\alias{G.nn.garrotte}
\title{Group nn Garrotte}
\usage{
G.nn.garrotte(x, y, g.index, p, normalized = TRUE, gram = TRUE)
}
\arguments{
\item{x}{design matrix for linear regression}

\item{y}{response variable}

\item{g.index}{a vector containing group information for each variable in \code{x}}

\item{p}{a vector containing the number of variables in the same group}

\item{normalized}{a logical value, default is "TRUE". If it is TRUE, each column of \code{x} will be scaled}

\item{gram}{a logical value, default is "TRUE". If it is TRUE, the gram-Schmidit orthonormalization will be applied for each group}
}
\value{
a list containing the origin data \code{x}, \code{y}, \code{g.index}, \code{p}, and the d-matrix \code{d}, estimated parameters \code{betas}, solution path \code{path}
}
\description{
Group non-negative Garrotte algorithm using R
}
\examples{
\dontrun{
  library(MASS)
  x<-mvrnorm(200,mu = rep(0,8),Sigma = diag(rep(1,8)))
  eps<-rnorm(200,0,1)
  beta<-matrix(c(.1,.2,.3,.3,.5,.9,.4,.4))
  y<-x\%*\%beta+eps
  g.index<-c(1,1,2,2,2,2,3,3)
  p<-c(2,4,2)
  fit<-G.nn.garrotte(x,y,g.index,p)
  fit
}
}
