\name{SA23204161-package}
\alias{SA23204161-package}
\alias{SA23204161}
\docType{package}
\title{
  R package for statistical computing of USTC
}
\description{
  Sets of functions of homework of statistical computing
}
\author{
Hu Bosen

Maintainer: Hu Bosen <hubsmx@mail.ustc.edu.cn>
}
\keyword{ package }
