% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SNSMS_Cpp}
\alias{SNSMS_Cpp}
\title{SNSMS for RCpp version}
\usage{
SNSMS_Cpp(L, x, theta, m, alpha, Tm, d)
}
\arguments{
\item{L}{a d-dimension object function}

\item{x}{a d-dimension time series data, expressed as a matrix}

\item{theta}{parameter estimates for the first m datas}

\item{m}{the number of observed data}

\item{alpha}{significance level}

\item{Tm}{the ratio of the monitoring horizon to the size of the training sample}

\item{d}{the dimension of data}
}
\value{
a random sample of size \code{k}
}
\description{
Self-normalized sequential change-point detection using RCpp
}
\examples{
\dontrun{
  L<-function(x,mu){(x-mu)^2}
  x<-matrix(c(rnorm(200,0,0.1),rnorm(1800,2,0.1)),nrow = 1)
  m<-180
  theta<-mean(x[1:m])
  alpha<-0.05
  Tm<-10
  d<-1
  k<-SNSMS_Cpp(L,x,theta,m,alpha,Tm,d)
  plot(x,type="l")
  abline(v=m+k,col="red")
}
\dontrun{
  library(MASS)
  L<-function(x,mu){(x-mu)^2}
  x<-cbind(t(mvrnorm(200,c(0,0),diag(c(1,1)))),t(mvrnorm(1800,c(1,2),diag(c(1,1)))))
  m<-180
  theta<-rowMeans(x[,1:m])
  alpha<-0.05
  Tm<-10
  d<-2
  k<-SNSMS_Cpp(L,x,theta,m,alpha,Tm,d)
}
}
